requirements.txt
streamlit
snowflake-snowpark-python
altair
pandas